Pass 3 Task 3: extract GhostPlaybackState to own file
Move LightPlaybackState, GhostPlaybackState, InterpolationResult from
nested types inside ParsekFlight to top-level types in
GhostPlaybackState.cs. Update 15 qualified references in ParsekKSC.
Breaks ParsekKSC->ParsekFlight type coupling. Zero logic changes.
