+++ Source/Parsek/ParsekFlight.cs
@@ -38,73 +38,6 @@ namespace Parsek
         private Recording previewRecording;
         internal int lastPlaybackIndex = 0;
 
-        // Per-ghost playback state (bundled to prevent dict-sync bugs)
-        internal class LightPlaybackState
-        {
-            public bool isOn;
-            public bool blinkEnabled;
-            public float blinkRateHz = 1f;
-        }
-
-        internal class GhostPlaybackState
-        {
-            public GameObject ghost;
-            public List<Material> materials;
-            public int playbackIndex;
-            public int partEventIndex;
-            public int loopCycleIndex = -1;
-            public Dictionary<uint, List<uint>> partTree;
-            public Dictionary<uint, ParachuteGhostInfo> parachuteInfos;
-            public Dictionary<uint, JettisonGhostInfo> jettisonInfos;
-            public Dictionary<ulong, EngineGhostInfo> engineInfos; // key = EncodeEngineKey(pid, moduleIndex)
-            public Dictionary<ulong, RcsGhostInfo> rcsInfos;   // separate from engineInfos — keys can overlap for same part
-            public Dictionary<ulong, RoboticGhostInfo> roboticInfos; // key = EncodeEngineKey(pid, moduleIndex)
-            public Dictionary<uint, DeployableGhostInfo> deployableInfos;
-            public Dictionary<uint, HeatGhostInfo> heatInfos;
-            public Dictionary<uint, LightGhostInfo> lightInfos;
-            public Dictionary<uint, LightPlaybackState> lightPlaybackStates;
-            public Dictionary<uint, List<ColorChangerGhostInfo>> colorChangerInfos;
-            public Dictionary<uint, FairingGhostInfo> fairingInfos;
-            public Dictionary<uint, GameObject> fakeCanopies;
-            public ReentryFxInfo reentryFxInfo;
-            public MaterialPropertyBlock reentryMpb; // per-ghost to avoid shared-state bugs with overlapping ghosts
-            public bool explosionFired;
-            public bool pauseHidden;
-            public bool rcsSuppressed;
-            public Transform cameraPivot; // child of ghost; centroid of active parts — camera targets this
-            public Vector3 lastInterpolatedVelocity;
-            public string lastInterpolatedBodyName;
-            public double lastInterpolatedAltitude;
-
-            public void SetInterpolated(InterpolationResult r)
-            {
-                lastInterpolatedVelocity = r.velocity;
-                lastInterpolatedBodyName = r.bodyName;
-                lastInterpolatedAltitude = r.altitude;
-            }
-        }
-
-        internal struct InterpolationResult
-        {
-            public Vector3 velocity;
-            public string bodyName;
-            public double altitude;
-
-            public static readonly InterpolationResult Zero = new InterpolationResult
-            {
-                velocity = Vector3.zero,
-                bodyName = null,
-                altitude = 0
-            };
-
-            public InterpolationResult(Vector3 vel, string body, double alt)
-            {
-                velocity = vel;
-                bodyName = body;
-                altitude = alt;
-            }
-        }
-
         private Dictionary<int, GhostPlaybackState> ghostStates = new Dictionary<int, GhostPlaybackState>();
         private readonly MaterialPropertyBlock reentryShellMpb = new MaterialPropertyBlock();
         private readonly List<GameObject> activeExplosions = new List<GameObject>();
